requirements.txt
mcp
playwright
google-genai
python-dotenv
httpx
pyperclip
